feat: add product card styles
