+++ src/shared/router.ts
@@ -1,8 +1,7 @@
 import { getLoginPageView } from "../pages/loginPage/loginPageView";
 import { MainPageController } from "../pages/mainPage/MainPageController";
 import RegistrationPageView from "../pages/registrationPage/registrationView";
-import { cardProductView } from "../widgets/cardProduct/cardProductView";
-import { categoryNavigationController } from "../widgets/categoryNavigation/categoryNavigationController";
+
 
 type ControllerFunction = () => void;
 
@@ -22,9 +21,6 @@ async function homeController(): Promise<void> {
   if (appContainer) {
     appContainer.innerHTML = ''
     appContainer.append(await MainPageController());
-    appContainer.append(await categoryNavigationController())
-    appContainer.append(await cardProductView())
-    
   }
 }
