+++ src/types/interfaces/widgets/ProductCard.ts
@@ -1,0 +1,3 @@
+export interface ProductCard {
+  [key: string]: string;
+}
