+++ src/widgets/cardProduct/cardProductView.ts
@@ -1,12 +1,11 @@
+import { ProductCard } from "../../types/interfaces/widgets/ProductCard";
 import { createElement } from "../../utils/abstract";
 
-interface ProductCard {
-  [key: string]: string;
-}
-
+// TO DO: add this to call function
 const productCard: ProductCard = {
   'img': 'link-to-image',
   'name': 'Name',
+  'info': 'Super puper cool plants for your garden',
   'price': 'Price',
   'sale': 'Sale-Price'
 }
@@ -15,22 +14,23 @@ export async function cardProductView(): Promise<HTMLElement> {
   const cardProductWrapper = createElement('div', ['product-card__wrapper']);
 
   cardProductWrapper.innerHTML = ` 
-    <div class="product-image">
-      <a href="">
+    <div class="product-card__image">
+      <a href="${productCard.link}">
         <img src="${productCard.img}">
-        <div class="shadow"></div>
+        <div class="product-card__shadow"></div>
       </a>
-      <a class="detail-link" href="" title="fast show"></a>
-      <div class="actions">
-        <div class="actions-btn">
-          <a class="add-to-cart" href="" title="add to cart"></a>
-          <a class="add-to-wishlist" href="" title="add to favorite"></a>
+      <a class="product-card__detail-link" href="${productCard.link}"></a>
+      <div class="product-card__actions">
+        <div class="product-card__actions-btn">
+          <a class="product-card__add-to-cart" href="" title="add to cart"></a>
+          <a class="product-card__add-to-wishlist" href="" title="add to favorite"></a>
         </div>
       </div> 
     </div>
-    <div class="product-list">
+    <div class="product-card__list">
       <h3>${productCard.name}</h3>
-      <div class="price"><span class="price sale">${productCard.sale}</span> ${productCard.price}</div>
+      <p class="product-card__sub-info">${productCard.info}</p>
+      <div class="product-card__price"><span class="product-card__price product-card__sale">${productCard.sale}</span> ${productCard.price}</div>
     </div>`
 
   return cardProductWrapper;
